requirements.txt
Mastodon.py ~= 1.3.1
Pillow ~= 5.3.0
